Move basicMapGenerator to own header file
